# incubator-backend/openapi.yaml
openapi: 3.0.3
info:
  title: IT Youth Talent Incubator API
  description: |
    # IT Youth Talent Incubator Backend API
    
    Welcome to the comprehensive API documentation for the IT Youth Talent Incubator platform.
    This platform connects talented IT students in Ghana with companies seeking skilled professionals.
    
    ## Features
    - **Authentication & Authorization**: Secure JWT-based authentication with role-based access control
    - **User Management**: Support for Students, Companies, and Admins with distinct profiles
    - **Job Posting & Management**: Companies can post, manage, and track jobs
    - **Application System**: Students can apply to jobs and track application status
    - **Admin Panel**: Comprehensive admin tools for user approval and platform management
    
    ## Authentication
    Most endpoints require authentication via JWT tokens. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your_token>
    ```
    
    ## Rate Limiting
    API requests are rate-limited to prevent abuse:
    - General endpoints: 100 requests per 15 minutes
    - Authentication endpoints: 3-5 requests per 15 minutes
    
    ## Pagination
    List endpoints support pagination with the following query parameters:
    - `page`: Page number (default: 1)
    - `limit`: Items per page (default: 10, max: 100)
    
  version: 1.0.0
  contact:
    name: IT Youth Talent Incubator
    email: support@itfy-incubator.com
  license:
...
servers:
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login or registration
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Operation successful"
        data:
          type: object
...
    Company:
      type: object
      properties:
        _id:
          type: string
          format: objectid
        user:
          $ref: '#/components/schemas/User'
        name:
          type: string
          example: "Tech Innovations Ghana"
        description:
          type: string
          example: "Leading software development company in West Africa"
        industry:
          type: string
          example: "Information Technology"
        website:
          type: string
          example: "https://techinnovations.com"
        social_links:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
                example: "LinkedIn"
              url:
                type: string
                example: "https://linkedin.com/company/techinnovations"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
          
    # ==================
    # Job Schemas
    # ==================
...
  responses:
    UnauthorizedError:
      description: Authentication token is missing or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            message: "Unauthorized"
            error: "You are not authorized to perform this action"
...
paths:
...
  /users/me:
    get:
      tags:
        - Users
      summary: Get my profile
      description: Retrieve the authenticated user's full profile
      operationId: getMyProfile
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Profile retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    oneOf:
                      - $ref: '#/components/schemas/Student'
                      - $ref: '#/components/schemas/Company'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

    put:
      tags:
        - Users
      summary: Update my profile
      description: Update the authenticated user's profile
      operationId: updateMyProfile
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Profile fields to update (varies by role)
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

    delete:
      tags:
        - Users
      summary: Delete my account
      description: Soft delete the authenticated user's account
      operationId: deleteMyAccount
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Account deleted successfully
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /users/me/photo:
    post:
      tags:
        - Users
      summary: Upload profile photo
      description: Upload or update profile photo
      operationId: uploadPhoto
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                photo:
                  type: string
                  format: binary
                  description: Profile photo (max 5MB, images only)
      responses:
        '200':
          description: Photo uploaded successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      photo_url:
                        type: string
        '400':
          description: Invalid file type or size
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /users:
    get:
      tags:
        - Users
      summary: List all users (Admin only)
      description: Retrieve paginated list of all users
      operationId: listUsers
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/RoleParam'
        - $ref: '#/components/parameters/StatusParam'
      responses:
        '200':
          description: Users retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /users/search:
    get:
      tags:
        - Users
      summary: Search users (Admin only)
      description: Search users by email, name, or other criteria
      operationId: searchUsers
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/SearchParam'
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'

  # ==================
  # Student Endpoints
  # ==================
  /users/students:
    get:
      tags:
        - Students
      summary: List all students
      description: |
        Retrieve paginated list of student profiles.
        - Public access: Shows approved students only
        - Admin access: Shows all students with optional filters
      operationId: listStudents
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/StatusParam'
      responses:
        '200':
          description: Students retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Student'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
        '400':
          $ref: '#/components/responses/ValidationError'
...
  /users/companies:
    get:
      tags:
        - Companies
      summary: List all companies
      description: |
        Retrieve paginated list of company profiles.
        - Public access: Shows approved companies only
        - Admin access: Shows all companies with optional filters
      operationId: listCompanies
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/StatusParam'
      responses:
        '200':
          description: Companies retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Company'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
        '400':
          $ref: '#/components/responses/ValidationError'
...
  /users/admins:
    get:
      tags:
        - Users
      summary: List all admins (Super Admin only)
      description: Retrieve list of admin users
      operationId: listAdmins
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Admins retrieved successfully
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /users/stats:
    get:
      tags:
        - Users
      summary: Get user statistics (Admin only)
      description: Retrieve platform user statistics
      operationId: getUserStats
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Statistics retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      totalUsers:
                        type: integer
                      activeUsers:
                        type: integer
                      pendingApprovals:
                        type: integer
                      studentCount:
                        type: integer
                      companyCount:
                        type: integer
                      adminCount:
                        type: integer
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /users/{id}/approve:
    post:
      tags:
        - Users
      summary: Approve user account (Admin only)
      description: Approve a pending user registration
      operationId: approveUser
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: User ID to approve
      responses:
        '200':
          description: User approved successfully
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /users/{id}/reject:
    post:
      tags:
        - Users
      summary: Reject user account (Admin only)
      description: Reject a pending user registration
      operationId: rejectUser
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: User ID to reject
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
                  description: Reason for rejection
      responses:
        '200':
          description: User rejected successfully
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /users/bulk-approve:
    post:
      tags:
        - Users
      summary: Bulk approve users (Admin only)
      description: Approve multiple users at once
      operationId: bulkApprove
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userIds:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: Users approved successfully
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /users/{id}/activate:
    put:
      tags:
        - Users
      summary: Activate user account (Admin only)
      description: Activate a deactivated user account
      operationId: activateUser
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User activated successfully
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /users/{id}/deactivate:
    put:
      tags:
        - Users
      summary: Deactivate user account (Admin only)
      description: Temporarily deactivate a user account
      operationId: deactivateUser
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User deactivated successfully
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /users/{id}:
    delete:
      tags:
        - Users
      summary: Delete user (Admin only)
      description: Soft delete a user account
      operationId: deleteUser
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User deleted successfully
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /users/{id}/force:
    delete:
      tags:
        - Users
      summary: Force delete user (Admin only)
      description: Permanently delete a user account
      operationId: forceDeleteUser
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User permanently deleted
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /users/{id}/restore:
    post:
      tags:
        - Users
      summary: Restore deleted user (Admin only)
      description: Restore a soft-deleted user account
      operationId: restoreUser
      security:
        - BearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User restored successfully
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'

  # ==================
  # Job Endpoints
  # ==================
...